requirements.txt
alembic==1.13.0
colorama==0.4.6
fastapi==0.104.1
setuptools==65.5.0
SQLAlchemy==2.0.23
uvicorn==0.24.0
passlib==1.7.4
python-jose==3.3.0
psycopg2-binary==2.9.9
email-validator==2.1.0
cryptography==41.0.7